fix out of range data

+++ database/migrations/2025_05_14_062930_create_order_parts_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('order_parts', function (Blueprint $table) {
             $table->id();
             $table->foreignId('part_id')->constrained('parts')->onDelete('cascade');
-            $table->float('part_price');
+            $table->decimal('part_price', 12, 2);
             $table->integer('qty');
             $table->timestamps();
         });
