+++ database/migrations/2025_05_14_063008_create_order_details_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('order_details', function (Blueprint $table) {
             $table->id();
             $table->string('no_invoice', 50);
-            $table->float('labor_cost_service');
+            $table->decimal('labor_cost_service', 12, 2);
             $table->timestamps();
         });
     }
